functions/api-round-attempts/python/requirements.txt
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.1
deprecation==2.1.0
Flask==2.2.2
functions-framework==3.2.0
google-api-core==2.10.1
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
googleapis-common-protos==1.56.4
grpcio==1.49.0
grpcio-status==1.49.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
